pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = ""
authors = [
    { name = "ankitraj20616", email = "ankitrajthakur803@gmail.com" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
langchain = "^1.2.0"
langchain-pinecone = "^0.2.13"
python-dotenv = "^1.2.1"
langchain-community = "^0.4.1"
pypdf = "^6.5.0"
langchain-text-splitters = "^1.1.0"
google-genai = "^1.56.0"
langchain-google-genai = "^4.1.2"
langchain-openai = "^1.1.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
